chore: reply code review
move hooks to web3-shared-evm
